LipSync with Visimes Animations implemented
To test LipSync with an Audio File, go to Audio Source in the Face Object and cross "Play on Awake". To Use It with a microphone just activate the SalsaMicInput Component and cross "Link with SALSA". This overrides the selected Audio Source. When using a microphone, some adjustments in the SALSA > Settings should be made to improve dynamics.

Add more Visimes as Animations to improve the smoothness in the transitions.